+++ src/frontend/tests/end-to-end/autoLogin.spec.ts
