+++ src/frontend/tests/end-to-end/codeAreaModalComponent.spec.ts
@@ -20,7 +20,9 @@ test("CodeAreaModalComponent", async ({ page }) => {
     modalCount = await page.getByTestId("modal-title")?.count();
   }
 
-  await page.waitForTimeout(1000);
+  await page.waitForSelector('[data-testid="blank-flow"]', {
+    timeout: 30000,
+  });
 
   await page.getByTestId("blank-flow").click();
