+++ src/frontend/tests/end-to-end/langflowShortcuts.spec.ts
@@ -67,11 +67,9 @@ test("LangflowShortcuts", async ({ page }) => {
     expect(false).toBeTruthy();
   }
 
-  await page
-    .locator(
-      '//*[@id="react-flow-id"]/div[1]/div[1]/div[1]/div/div[2]/div[2]/div/div[1]/div/div[1]/div/div/div[1]',
-    )
-    .click();
+  const ollamaTitleElement = await page.getByTestId("title-Ollama").last();
+
+  await ollamaTitleElement.click();
   await page.keyboard.press("Backspace");
 
   numberOfNodes = await page.getByTestId("title-Ollama")?.count();
@@ -90,11 +88,7 @@ test("LangflowShortcuts", async ({ page }) => {
     expect(false).toBeTruthy();
   }
 
-  await page
-    .locator(
-      '//*[@id="react-flow-id"]/div[1]/div[1]/div[1]/div/div[2]/div[2]/div/div[1]/div/div[1]/div/div/div[1]',
-    )
-    .click();
+  await ollamaTitleElement.click();
   await page.keyboard.press("Backspace");
 
   await page.getByTestId("title-Ollama").click();
